Splitting myFares into smaller tests
